refactor: solve path error
