+++ src/components/user/User.tsx
@@ -3,17 +3,17 @@ import UserCard from './components/user-card/UserCard';
 import UserInteractions from './components/user-interactions/UserInteractions';
 import { ReactDuo } from '../../utils/react';
 import { useEffect, useState, useContext } from 'react';
-import { UserEntity, UserWithToken } from '../../../ForFable-Domain';
+import { UserEntity } from '../../../ForFable-Domain';
 import { LanguageContext } from '../../contexts/LanguageContext';
-import { ServicesContext } from 'src/contexts/ServicesContext';
+import { ServicesContext } from '../../contexts/ServicesContext';
 
 interface UserProps {
-  userDuo: ReactDuo<UserWithToken|undefined>
+  requestUserDuo: ReactDuo<UserEntity|null>
   searchUserId: number
 }
 
-const User: React.FC<UserProps> = ({userDuo, searchUserId }) => {
-  const [userData] = userDuo
+const User: React.FC<UserProps> = ({requestUserDuo, searchUserId }) => {
+  const [userData] = requestUserDuo
   const searchUserDuo = useState<UserEntity|null>()
   const [searchUser, setSearchUser] = searchUserDuo
   const [lang] = useContext(LanguageContext)
@@ -24,19 +24,23 @@ const User: React.FC<UserProps> = ({userDuo, searchUserId }) => {
 
   useEffect(() => {
     const loadUser = async () => {
+        if(searchUserId==userData?.id){
+            setSearchUser(userData)
+            return
+        }
         const response = await UsersService.show(searchUserId)
         if (response.state=='Sucess') {
             setSearchUser(response.data)
         }
     }
     loadUser()
-  }, [searchUserId, setSearchUser, UsersService])
+  }, [searchUserId, setSearchUser, UsersService, userData])
 
   if(!searchUser) { return <div> {lang.Loading}... </div> }
 
   return (
     <div className='block'>
-      <UserCard userDuo={(searchUserDuo[0]?.id == userData?.id ? userDuo : searchUserDuo) as ReactDuo<UserEntity>} isUser={isUser}/>
+      <UserCard userDuo={(searchUserId === userData?.id ? requestUserDuo : searchUserDuo) as ReactDuo<UserEntity>} isUser={isUser}/>
       <UserInteractions userId={searchUser.id} />
     </div>
   );
