+++ src/components/user/components/user-card/components/color-pickers/ColorPickers.tsx
@@ -1,6 +1,6 @@
 import { HexColorPicker } from 'react-colorful';
 import './ColorPickers.css'
-import { ReactDuo } from '../../../../../../../src/utils/react';
+import { ReactDuo } from '../../../../../../utils/react';
 
 
 interface ColorPickersProps {
